worked on datafeed file generation logging and plugin upgrade process

+++ includes/class-shareasale-wc-tracker-uninstaller.php
@@ -6,14 +6,20 @@ if ( ! defined( 'WPINC' ) ) {
 class ShareASale_WC_Tracker_Uninstaller {
 
 	public static function uninstall() {
+		if ( ! current_user_can( 'activate_plugins' ) ) {
+			return;
+		}
+
 		global $wpdb;
-		$logs_table = $wpdb->prefix . 'shareasale_wc_tracker_logs';
-		//drop logs table
-		$query = 'DROP TABLE ' . $logs_table;
+		$logs_table     = $wpdb->prefix . 'shareasale_wc_tracker_logs';
+		$datafeed_table = $wpdb->prefix . 'shareasale_wc_tracker_datafeeds';
+		//drop tables
+		$query = 'DROP TABLE ' . $logs_table . ', ' . $datafeed_table;
 		$wpdb->query( $query );
 		//remove settings
 		unregister_setting( 'shareasale_wc_tracker_options', 'shareasale_wc_tracker_options' );
 		delete_option( 'shareasale_wc_tracker_options' );
+		delete_option( 'shareasale_wc_tracker_version' );
 	}
 
 	public static function disable() {
